Move all UnionFind classes to new subpackage union_find
